refactor(redis): move makeIoredisClients to /ioredis subpath (Copilot review)
Copilot PR #102 review (and prior Codex multi-agent review): the main
public-API entry leaks the ioredis Redis type into dist/clients.d.mts.
Even with ioredis as an optional peer, TypeScript module resolution
needs the type files installed; consumers importing the main entry but
not using makeIoredisClients still get "Cannot find module 'ioredis'".

Fix per Copilot suggestion (a): isolate the vendor-specific helper to
@o3co/auth-provider-redis/ioredis subpath. Same convention as
@o3co/auth-provider-core/testing (Phase 9 PR α).

- packages/redis/src/clients.mts → src/ioredis.mts (renamed, no logic change)
- packages/redis/package.json: + "./ioredis" exports subpath
- packages/redis/src/index.mts: drop the makeIoredisClients re-export;
  add comment pointing at the subpath
- 8 test files: import { makeIoredisClients } from "../src/ioredis.mjs"
  instead of "../src/index.mjs"

Main entry now stays vendor-agnostic. Future per-vendor wrappers
(e.g. node-redis) follow the same /<vendor> subpath pattern.
