+++ packages/redis/__tests__/redis.sessionRPRegistry.test.mts
@@ -17,7 +17,7 @@
 import Redis from "ioredis";
 import { GenericContainer, type StartedTestContainer } from "testcontainers";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
-import { makeIoredisClients } from "../src/index.mjs";
+import { makeIoredisClients } from "../src/ioredis.mjs";
 import { createRedisSessionRPRegistry } from "../src/sessionRPRegistry.mjs";
 import { runSessionRPRegistryContract } from "./sessionRPRegistry.contract.mjs";
