+++ packages/redis/__tests__/redisSessionStoresModule.test.mts
@@ -19,7 +19,7 @@ import { makeValidCoreConfig } from "@o3co/auth-provider-core/testing";
 import Redis from "ioredis";
 import { GenericContainer, type StartedTestContainer } from "testcontainers";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
-import { makeIoredisClients } from "../src/index.mjs";
+import { makeIoredisClients } from "../src/ioredis.mjs";
 import { redisSessionStoresModule } from "../src/modules/redisSessionStores.mjs";
 
 let container: StartedTestContainer;
