+++ packages/redis/__tests__/refresh-token-family-wiring.test.mts
@@ -24,7 +24,8 @@ import { makeValidCoreConfig } from "@o3co/auth-provider-core/testing";
 import Redis from "ioredis";
 import { GenericContainer, type StartedTestContainer } from "testcontainers";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
-import { makeIoredisClients, redisRefreshTokenFamilyStoreModule } from "../src/index.mjs";
+import { redisRefreshTokenFamilyStoreModule } from "../src/index.mjs";
+import { makeIoredisClients } from "../src/ioredis.mjs";
 
 let container: StartedTestContainer;
 let client: Redis;
