+++ packages/redis/__tests__/wiring.test.mts
@@ -23,11 +23,8 @@ import { makeValidCoreConfig } from "@o3co/auth-provider-core/testing";
 import Redis from "ioredis";
 import { GenericContainer, type StartedTestContainer } from "testcontainers";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
-import {
-	makeIoredisClients,
-	redisChallengeStoreModule,
-	redisReplaySeenSetModule,
-} from "../src/index.mjs";
+import { redisChallengeStoreModule, redisReplaySeenSetModule } from "../src/index.mjs";
+import { makeIoredisClients } from "../src/ioredis.mjs";
 
 let container: StartedTestContainer;
 let client: Redis;
